+++ src/main/java/bridge/domain/BridgeMaker.java
@@ -1,5 +1,6 @@
-package bridge;
+package bridge.domain;
 
+import bridge.domain.generator.BridgeNumberGenerator;
 import java.util.List;
 
 /**
